run_tests: use same everywhere and clean test dir on success

+++ plugins/plugin-check-amt-montreal/run_tests.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
+rm -rf env
 virtualenv env
 source env/bin/activate
 pip install -I nose
 pip install -r requirements.txt
-nosetests
-rm -rf env
+nosetests && rm -rf env
